Moving form logic to new Base Form class
